# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: API de Gestion de Solicitudes Academicas
  version: 1.0.0
  description: API REST disenada desde el dominio para gestionar usuarios y solicitudes academicas.
servers:
  - url: http://localhost:8081
    description: Desarrollo local
tags:
  - name: Solicitudes
    description: Operaciones del agregado Solicitud
  - name: Usuarios
    description: Operaciones del agregado Usuario
paths:
  /api/v1/solicitudes:
    post:
      tags: [Solicitudes]
      summary: Crear solicitud
      description: Registra una nueva solicitud academica.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrearSolicitudRequest'
      responses:
        '201':
          description: Solicitud creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolicitudDetalleResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags: [Solicitudes]
      summary: Listar solicitudes
      description: Retorna todas las solicitudes o filtra por estado con `estado`.
      parameters:
        - in: query
          name: estado
          required: false
          schema:
            type: string
            enum: [REGISTRADA, CLASIFICADA, EN_ATENCION, ATENDIDA, CERRADA]
      responses:
        '200':
          description: Solicitudes consultadas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SolicitudResumenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/solicitudes/{codigo}:
    get:
      tags: [Solicitudes]
      summary: Consultar solicitud por codigo
      description: Retorna el detalle de una solicitud.
      parameters:
        - $ref: '#/components/parameters/CodigoSolicitud'
      responses:
        '200':
          description: Solicitud encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolicitudDetalleResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/solicitudes/{codigo}/responsable:
    put:
      tags: [Solicitudes]
      summary: Asignar responsable
      description: Asigna un funcionario responsable a una solicitud.
      parameters:
        - $ref: '#/components/parameters/CodigoSolicitud'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AsignarResponsableRequest'
      responses:
        '200':
          description: Responsable asignado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolicitudDetalleResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/solicitudes/{codigo}/clasificacion:
    post:
      tags: [Solicitudes]
      summary: Clasificar solicitud
      description: Clasifica una solicitud registrada y define su tipo.
      parameters:
        - $ref: '#/components/parameters/CodigoSolicitud'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClasificarSolicitudRequest'
      responses:
        '200':
          description: Solicitud clasificada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolicitudDetalleResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/solicitudes/{codigo}/prioridad:
    post:
      tags: [Solicitudes]
      summary: Priorizar solicitud
      description: Asigna prioridad a una solicitud clasificada.
      parameters:
        - $ref: '#/components/parameters/CodigoSolicitud'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PriorizarSolicitudRequest'
      responses:
        '200':
          description: Solicitud priorizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolicitudDetalleResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/solicitudes/{codigo}/atencion/inicio:
    post:
      tags: [Solicitudes]
      summary: Iniciar atencion
      description: Cambia la solicitud al estado EN_ATENCION.
      parameters:
        - $ref: '#/components/parameters/CodigoSolicitud'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IniciarAtencionRequest'
      responses:
        '200':
          description: Atencion iniciada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolicitudDetalleResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/solicitudes/{codigo}/atencion/finalizacion:
    post:
      tags: [Solicitudes]
      summary: Finalizar atencion
      description: Marca la solicitud como ATENDIDA.
      parameters:
        - $ref: '#/components/parameters/CodigoSolicitud'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MarcarAtendidaRequest'
      responses:
        '200':
          description: Solicitud atendida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolicitudDetalleResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/solicitudes/{codigo}/cierre:
    post:
      tags: [Solicitudes]
      summary: Cerrar solicitud
      description: Cierra una solicitud atendida con observacion administrativa.
      parameters:
        - $ref: '#/components/parameters/CodigoSolicitud'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CerrarSolicitudRequest'
      responses:
        '200':
          description: Solicitud cerrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolicitudDetalleResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/usuarios:
    post:
      tags: [Usuarios]
      summary: Crear usuario
      description: Registra un nuevo usuario.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrearUsuarioRequest'
      responses:
        '201':
          description: Usuario creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioDetalleResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags: [Usuarios]
      summary: Listar usuarios
      description: Retorna la lista resumida de usuarios registrados.
      responses:
        '200':
          description: Usuarios consultados
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UsuarioResumenResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/usuarios/{id}:
    get:
      tags: [Usuarios]
      summary: Consultar usuario por id
      description: Retorna el detalle de un usuario.
      parameters:
        - $ref: '#/components/parameters/IdUsuario'
      responses:
        '200':
          description: Usuario encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioDetalleResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/health:
    get:
      tags: [Usuarios]
      summary: Health check
      description: Verifica que el adaptador HTTP este disponible.
      responses:
        '200':
          description: Aplicacion disponible
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
components:
  parameters:
    CodigoSolicitud:
      name: codigo
      in: path
      required: true
      description: Codigo unico de la solicitud
      schema:
        type: string
    IdUsuario:
      name: id
      in: path
      required: true
      description: Identificador unico del usuario
      schema:
        type: string
  responses:
    BadRequest:
      description: Datos invalidos o regla de dominio incumplida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: El usuario no tiene permisos para ejecutar la operacion
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Estado del recurso incompatible con la operacion
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Error interno del servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    CrearSolicitudRequest:
      type: object
      required: [estudianteId, canal, tipo, descripcion]
      properties:
        estudianteId:
          type: string
        canal:
          type: string
          enum: [CSU, CORREO, SAC, TELEFONICO]
        tipo:
          type: string
          enum: [REGISTRO_ASIGNATURA, HOMOLOGACION, CANCELACION, SOLICITUD_CUPO, CONSULTA_ACADEMICA]
        descripcion:
          type: string
          minLength: 20
          maxLength: 1000
    AsignarResponsableRequest:
      type: object
      required: [funcionarioId]
      properties:
        funcionarioId:
          type: string
    ClasificarSolicitudRequest:
      type: object
      required: [funcionarioId, tipo]
      properties:
        funcionarioId:
          type: string
        tipo:
          type: string
          enum: [REGISTRO_ASIGNATURA, HOMOLOGACION, CANCELACION, SOLICITUD_CUPO, CONSULTA_ACADEMICA]
    PriorizarSolicitudRequest:
      type: object
      required: [funcionarioId, nivel, justificacion]
      properties:
        funcionarioId:
          type: string
        nivel:
          type: string
          enum: [BAJA, MEDIA, ALTA, CRITICA]
        justificacion:
          type: string
          minLength: 5
          maxLength: 500
    IniciarAtencionRequest:
      type: object
      required: [funcionarioId]
      properties:
        funcionarioId:
          type: string
    MarcarAtendidaRequest:
      type: object
      required: [funcionarioId]
      properties:
        funcionarioId:
          type: string
    CerrarSolicitudRequest:
      type: object
      required: [administradorId, observacion]
      properties:
        administradorId:
          type: string
        observacion:
          type: string
          minLength: 10
          maxLength: 500
    CrearUsuarioRequest:
      type: object
      required: [nombre, email, tipo]
      properties:
        nombre:
          type: string
          minLength: 3
          maxLength: 100
        email:
          type: string
          format: email
        tipo:
          type: string
          enum: [ESTUDIANTE, FUNCIONARIO, ADMINISTRADOR]
    HistorialResponse:
      type: object
      properties:
        accion:
          type: string
        responsable:
          type: string
        observacion:
          type: string
        fecha:
          type: string
          format: date-time
    SolicitudResumenResponse:
      type: object
      properties:
        codigo:
          type: string
        estudianteId:
          type: string
        estudianteNombre:
          type: string
        tipo:
          type: string
        estado:
          type: string
        prioridad:
          type: string
          nullable: true
    SolicitudDetalleResponse:
      type: object
      properties:
        codigo:
          type: string
        estudianteId:
          type: string
        estudianteNombre:
          type: string
        canal:
          type: string
        tipo:
          type: string
        descripcion:
          type: string
        estado:
          type: string
        prioridad:
          type: string
          nullable: true
        historial:
          type: array
          items:
            $ref: '#/components/schemas/HistorialResponse'
    UsuarioResumenResponse:
      type: object
      properties:
        id:
          type: string
        nombre:
          type: string
        tipo:
          type: string
        estado:
          type: string
    UsuarioDetalleResponse:
      type: object
      properties:
        id:
          type: string
        nombre:
          type: string
        email:
          type: string
        tipo:
          type: string
        estado:
          type: string
    ErrorResponse:
      type: object
      properties:
        mensaje:
          type: string
        codigo:
          type: string
        timestamp:
          type: string
          format: date-time